Converting tj3 to use Tj3AppBase class.

+++ spec/Tj3Daemon_spec.rb
@@ -12,7 +12,7 @@
 #
 
 require 'rubygems'
-
+require 'support/StdIoWrapper'
 require 'taskjuggler/apps/Tj3Daemon'
 require 'taskjuggler/apps/Tj3Client'
 
@@ -20,61 +20,74 @@ class TaskJuggler
 
   describe Tj3Daemon do
 
-    Results = Struct.new(:returnValue, :stdOut, :stdErr)
+    include StdIoWrapper
 
-    it 'should be startable and stopable' do
-      start.should == 0
-      sleep 2
-      status.stdOut.should match /No projects registered/
-      stop.should == 0
-    end
-
-    def start
+    before(:each) do
       (f = File.new('taskjuggler.rc', 'w')).write(<<'EOT'
 _global:
   authKey: 'secret_key'
+  port: 0
+  _log:
+    outputLevel: 3
+    logLevel: 3
 EOT
                                                  )
       f.close
 
       if (pid = fork).nil?
+        at_exit { exit! }
         $stdout.reopen('stdout.log', 'w')
         $stderr.reopen('stderr.log', 'w')
-        Tj3Daemon.new.main(%w( --silent ))
+        res = stdIoWrapper do
+          Tj3Daemon.new.main(%w( --silent ))
+        end
+        raise "Failed to start tj3d: #{res.stdErr}" if res.returnValue != 0
+        exit!
+      else
+        # Wait for the daemon to get online.
+        i = 0
+        while !File.exists?('.tj3d.uri') && i < 10
+          sleep 0.5
+          i += 1
+        end
+        raise 'Daemon did not start properly' if i == 10
       end
       0
     end
 
-    def status
-      res = capture do
-        Tj3Client.new.main(%w( --unsafe --silent status ))
+    after(:each) do
+      res = stdIoWrapper do
+        Tj3Client.new.main(%w( --silent --unsafe terminate ))
       end
-      res.returnValue.should == 0
-      res.stdErr.should be_empty
-
-      res
-    end
-
-    def stop
-      res = Tj3Client.new.main(%w( --silent terminate ))
+      raise "tj3d termination failed: #{res.stdErr}" if res.returnValue != 0
+      i = 0
+      while File.exists?('.tj3d.uri') && i < 10
+        sleep 0.5
+        i += 1
+      end
+      raise "Daemon did not terminate properly" if i == 10
+      # Cleanup file system again.
       %w( taskjuggler.rc stdout.log stderr.log ).each do |file|
         File.delete(file)
       end
-      res
     end
 
-    def capture
-      oldStdOut = $stdout
-      oldStdErr = $stderr
-      $stdout = (out = StringIO.new)
-      $stderr = (err = StringIO.new)
-      begin
-        res = yield
-      ensure
-        $stdout = oldStdOut
-        $stderr = oldStdErr
+    it 'should be startable and stopable' do
+      res = stdIoWrapper do
+        Tj3Client.new.main(%w( --unsafe --silent status ))
       end
-      Results.new(res, out.string, err.string)
+      res.returnValue.should == 0
+      res.stdErr.should == ''
+      res.stdOut.should match /No projects registered/
+    end
+
+    it 'should be able to load a project' do
+      prj = 'project foo "Foo" 2011-03-14 +1d task "Foo"'
+      res = stdIoWrapper(prj) do
+        Tj3Client.new.main(%w( --unsafe --silent add . ))
+      end
+      res.returnValue.should == 0
+      res.stdErr.should match /Project foo loaded/
     end
 
   end
